refactor(ace): move package manager source under src
Move the Ace TypeScript package-manager implementation and bundled package anchors from tools/ace to src/Core.TypeScript/ace, keeping tools for setup/bootstrap edges.

Update the package bin, cross-verify workflow, build-gates doc, and cross-verification imports to follow the new source-owned path.
