+++ package.json
@@ -20,7 +20,7 @@
     "shadow": "bun tools/shadow/zeta-shadow.ts"
   },
   "bin": {
-    "ace": "tools/ace/ace.ts",
+    "ace": "src/Core.TypeScript/ace/ace.ts",
     "zeta-shadow": "tools/shadow/zeta-shadow.ts"
   },
   "devDependencies": {
