+++ src/Core.TypeScript/ace/index-signature.test.ts
