+++ src/Core.TypeScript/ace/index-signature.ts
