+++ src/Core.TypeScript/ace/lockfile.ts
