+++ src/Core.TypeScript/ace/package-hash.ts
