+++ src/Core.TypeScript/ace/packages/qsharp-reference-oracle-0.1.0.json
