+++ src/Core.TypeScript/ace/registry-publish.test.ts
