+++ src/Core.TypeScript/ace/registry-publish.ts
