+++ src/Core.TypeScript/ace/registry-remote.test.ts
