+++ src/Core.TypeScript/ace/registry-remote.ts
