+++ src/Core.TypeScript/ace/registry-revoke.test.ts
