+++ src/Core.TypeScript/ace/registry-revoke.ts
