+++ src/Core.TypeScript/ace/resolve.ts
