+++ src/Core.TypeScript/ace/setup-manifest.test.ts
