+++ src/Core.TypeScript/ace/setup-manifest.ts
