+++ src/Core.TypeScript/ace/signing.test.ts
