+++ src/Core.TypeScript/ace/signing.ts
