+++ src/Core.TypeScript/ace/solver.ts
