+++ tests/cross-verification/index-signature/cross-verify.ts
@@ -9,8 +9,8 @@
 // bytes -> expected_sig; sha256(SPKI-DER)[:16] -> expected_key_id), so no layer is a TS-vs-TS
 // tautology. Run from this directory: `bun cross-verify.ts`.
 import { createPrivateKey, createPublicKey } from "node:crypto";
-import { signIndex, verifyIndexSignature, canonicalIndexBytes, type IndexSignableContent } from "../../../tools/ace/index-signature.ts";
-import { type TrustEntry, type AceSignature } from "../../../tools/ace/signing.ts";
+import { signIndex, verifyIndexSignature, canonicalIndexBytes, type IndexSignableContent } from "../../../src/Core.TypeScript/ace/index-signature.ts";
+import { type TrustEntry, type AceSignature } from "../../../src/Core.TypeScript/ace/signing.ts";
 
 interface CanonVec { id: string; content: IndexSignableContent; expected_canonical_json: string; }
 interface EnvVec { id: string; private_pem: string; content: IndexSignableContent; expected_key_id: string; expected_sig: string; expected_canonical_json: string; }
